+++ test/ssh/test_dhparams.py
@@ -23,11 +23,16 @@ class TestSshDHParams(TestSshCases.TestSshClientBase):
     def test_real_gex(self):
         result = self.get_result('git.launchpad.net', 22)
         self.assertEqual(result.key_exchange.kex_algorithms, [SshKexAlgorithm.DIFFIE_HELLMAN_GROUP14_SHA1, ])
+        self.assertEqual(result.group_exchange.gex_algorithms, [
+            SshKexAlgorithm.DIFFIE_HELLMAN_GROUP_EXCHANGE_SHA256,
+            SshKexAlgorithm.DIFFIE_HELLMAN_GROUP_EXCHANGE_SHA1,
+        ])
         self.assertEqual(result.group_exchange.key_sizes, [2048, 3072, 4096, 6144, 7680, 8192])
         self.assertFalse(result.group_exchange.bounds_tolerated)
 
         result = self.get_result('github.com', 22)
         self.assertEqual(result.key_exchange, None)
+        self.assertEqual(result.group_exchange.gex_algorithms, [SshKexAlgorithm.DIFFIE_HELLMAN_GROUP_EXCHANGE_SHA256])
         self.assertEqual(result.group_exchange.key_sizes, [2048, 3072, 4096, 6144, 7680, 8192])
         self.assertTrue(result.group_exchange.bounds_tolerated)
 
@@ -38,5 +43,6 @@ class TestSshDHParams(TestSshCases.TestSshClientBase):
             SshKexAlgorithm.DIFFIE_HELLMAN_GROUP14_SHA256,
             SshKexAlgorithm.DIFFIE_HELLMAN_GROUP14_SHA1,
         ])
+        self.assertEqual(result.group_exchange.gex_algorithms, [SshKexAlgorithm.DIFFIE_HELLMAN_GROUP_EXCHANGE_SHA256])
         self.assertEqual(result.group_exchange.key_sizes, [2048, 3072, 4096, 6144, 7680, 8192])
         self.assertTrue(result.group_exchange.bounds_tolerated)